refactor: spring과 spark 이름 분리
