+++ src/main/java/chess/ConsoleChessApplication.java
@@ -2,7 +2,7 @@ package chess;
 
 import chess.controller.ChessController;
 
-public class Application {
+public class ConsoleChessApplication {
     public static void main(String[] args) {
         ChessController controller = new ChessController();
         controller.run();
