+++ src/main/java/chess/service/GameService.java
@@ -3,8 +3,8 @@ package chess.service;
 import java.util.List;
 
 import chess.database.GameStateGenerator;
-import chess.database.dao.BoardDao;
-import chess.database.dao.GameDao;
+import chess.database.vanillajdbc.dao.BoardDao;
+import chess.database.vanillajdbc.dao.GameDao;
 import chess.database.dto.BoardDto;
 import chess.database.dto.GameStateDto;
 import chess.database.dto.PointDto;
